fix: address sonar linq smell and close new-code coverage gaps
Agent-Logs-Url: https://github.com/schivei/GenDI/sessions/ee3e4b32-0ed3-4981-88b0-d0780c6dd5d3

+++ src/GenDI/ThreadIsolationPolicy.cs
@@ -15,15 +15,15 @@ public enum ThreadIsolationPolicy
     /// <summary>
     /// Enables thread-local caching with singleton registration semantics.
     /// </summary>
-    Singleton = ServiceLifetime.Singleton,
+    Singleton = (int)ServiceLifetime.Singleton,
 
     /// <summary>
     /// Enables thread-local caching with scoped registration semantics.
     /// </summary>
-    Scoped = ServiceLifetime.Scoped,
+    Scoped = (int)ServiceLifetime.Scoped,
 
     /// <summary>
     /// Enables thread-local caching with transient registration semantics.
     /// </summary>
-    Transient = ServiceLifetime.Transient,
+    Transient = (int)ServiceLifetime.Transient,
 }
